Updating scraper code -- still broken
