+++ src/apps/blockScrape/handle_block_scrape.cpp
@@ -121,6 +121,7 @@ bool COptions::scrape_blocks(void) {
         defaultQuitHandler(1);  // this does not quit, but only notifies the caller that the user quit blaze early
         EXIT_NOMSG(false);
     }
+    LOG_INFO("- <PROG> : Scraping ", cons.blazeCnt, " of ", cons.blazeCnt);
 
     if (!verbose) {
         cerr << '\r' << string_q(120, ' ') << '\r';
